Move battle.items, battle.tables to parent path
