--- nekoyume/battle/items/__init__.py
@@ -1,4 +1,0 @@
-from nekoyume.battle.items.item import Equipment, Item
-from nekoyume.battle.items.weapons import Weapon
-
-__all__ = ('Armor', 'Equipment', 'Item', 'Weapon')
