+++ nekoyume/items/foods.py
@@ -1,0 +1,8 @@
+from nekoyume.battle.enums import ItemType
+from nekoyume.battle.items.base import UseItem
+
+
+class Food(UseItem):
+    def __init__(self, name):
+        super().__init__(name)
+        self.type_ = ItemType.FOOD
