+++ nekoyume/items/heads.py
@@ -1,0 +1,8 @@
+from nekoyume.battle.enums import ItemType
+from nekoyume.battle.items.base import Equipment
+
+
+class Head(Equipment):
+    def __init__(self, name):
+        super().__init__(name)
+        self.type_ = ItemType.HEAD
