+++ nekoyume/tables/item_equips.tsv
