+++ nekoyume/tables/monster_skills.tsv
