+++ nekoyume/tables/monsters.tsv
