+++ nekoyume/tables/skills.tsv
