+++ setup.py
@@ -77,9 +77,11 @@ if __name__ == '__main__':
             ],
         },
         package_data={
-            'nekoyume': ['data/*', 'templates/*.html',
-                         'translations/*/LC_MESSAGES/*'],
-            'nekoyume.battle': ['tables/*.tsv'],
+            'nekoyume': [
+                'tables/*.tsv'
+                'templates/*.html',
+                'translations/*/LC_MESSAGES/*'
+                ],
         },
         entry_points={
             'console_scripts': [
